+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Fix FlawCollector to account for an empty acknowledgment affiliation (OSIDB-1195)
 
 ## [3.4.0] - 2023-08-14
 ### Added